src_datastore/cloud_functions/requirements.txt
google-cloud-firestore
google-cloud-datastore
google-cloud-storage
google-cloud-pubsub
protobuf==3.20.1